Make Client.join() not wait.
Consumer.pause() just set a flag requesting exit.
This flag was only provessed when queue items where processed
of when a timeout (0.5s) while witing for the queue was reached.

This patch adds a `flush()` method which wakes up the Consumer
by sending an empty message. It also ensures that `Client.flush()`
really forces the Consumer to flush.